[ledger] move the protocol module under ledger
