+++ apps/src/lib/node/mod.rs
@@ -1,3 +1,2 @@
 pub mod gossip;
 pub mod ledger;
-pub mod protocol;
